Split GUI audio dropdown tests
