+++ src/gui_app/tests.rs
@@ -12,6 +12,7 @@ use radiant::{
 use std::{fs, path::PathBuf, sync::mpsc};
 
 mod audio_settings_controls;
+mod audio_settings_dropdowns;
 mod shortcuts_context;
 mod status_bar;
 
@@ -689,223 +690,6 @@ fn sample_row_hit_target_survives_frame_refresh_between_press_and_release() {
         ))
     );
     assert!(!refreshed_hit_target.common().state.pressed);
-}
-
-#[test]
-fn audio_backend_dropdown_renders_expanded_host_options() {
-    let mut state = gui_state_for_span_tests();
-    state.audio_settings_error = None;
-    state.audio_backend_dropdown_open = true;
-    state.audio_hosts = vec![
-        super::AudioHostSummary {
-            id: String::from("wasapi"),
-            label: String::from("WASAPI"),
-            is_default: true,
-        },
-        super::AudioHostSummary {
-            id: String::from("asio"),
-            label: String::from("ASIO"),
-            is_default: false,
-        },
-    ];
-    let frame = radiant::runtime::UiSurface::new(super::audio_settings_popover(&state).into_node())
-        .frame(
-            Rect::from_min_size(Point::new(0.0, 0.0), Vector2::new(480.0, 360.0)),
-            &radiant::theme::ThemeTokens::default(),
-        );
-    let texts = frame
-        .paint_plan
-        .primitives
-        .iter()
-        .filter_map(|primitive| match primitive {
-            PaintPrimitive::Text(text) => Some(text.text.as_str().to_string()),
-            _ => None,
-        })
-        .collect::<Vec<_>>();
-
-    assert!(
-        texts.iter().any(|text| text == "System default"),
-        "{texts:?}"
-    );
-    assert!(
-        texts.iter().any(|text| text == "WASAPI (default)"),
-        "{texts:?}"
-    );
-    assert!(texts.iter().any(|text| text == "ASIO"), "{texts:?}");
-}
-
-#[test]
-fn audio_output_dropdown_renders_expanded_device_options() {
-    let mut state = gui_state_for_span_tests();
-    state.audio_settings_error = None;
-    state.audio_output_dropdown_open = true;
-    state.audio_devices = vec![super::AudioDeviceSummary {
-        host_id: String::from("asio"),
-        name: String::from("Studio Out"),
-        is_default: true,
-    }];
-    let frame = radiant::runtime::UiSurface::new(super::audio_settings_popover(&state).into_node())
-        .frame(
-            Rect::from_min_size(Point::new(0.0, 0.0), Vector2::new(480.0, 360.0)),
-            &radiant::theme::ThemeTokens::default(),
-        );
-    let texts = frame
-        .paint_plan
-        .primitives
-        .iter()
-        .filter_map(|primitive| match primitive {
-            PaintPrimitive::Text(text) => Some(text.text.as_str().to_string()),
-            _ => None,
-        })
-        .collect::<Vec<_>>();
-
-    assert!(texts.iter().any(|text| text == "Host default"), "{texts:?}");
-    assert!(
-        texts.iter().any(|text| text == "Studio Out (default)"),
-        "{texts:?}"
-    );
-}
-
-#[test]
-fn audio_sample_rate_dropdown_renders_expanded_rate_options() {
-    let mut state = gui_state_for_span_tests();
-    state.audio_settings_error = None;
-    state.audio_sample_rate_dropdown_open = true;
-    state.audio_sample_rates = vec![44_100, 48_000];
-    let frame = radiant::runtime::UiSurface::new(super::audio_settings_popover(&state).into_node())
-        .frame(
-            Rect::from_min_size(Point::new(0.0, 0.0), Vector2::new(480.0, 360.0)),
-            &radiant::theme::ThemeTokens::default(),
-        );
-    let texts = frame
-        .paint_plan
-        .primitives
-        .iter()
-        .filter_map(|primitive| match primitive {
-            PaintPrimitive::Text(text) => Some(text.text.as_str().to_string()),
-            _ => None,
-        })
-        .collect::<Vec<_>>();
-
-    assert!(
-        texts.iter().any(|text| text == "Device default"),
-        "{texts:?}"
-    );
-    assert!(texts.iter().any(|text| text == "44.1 kHz"), "{texts:?}");
-    assert!(texts.iter().any(|text| text == "48 kHz"), "{texts:?}");
-}
-
-#[test]
-fn audio_backend_dropdown_overlay_does_not_reflow_later_sections() {
-    let mut state = gui_state_for_span_tests();
-    state.audio_settings_error = None;
-    state.audio_hosts = vec![
-        super::AudioHostSummary {
-            id: String::from("wasapi"),
-            label: String::from("WASAPI"),
-            is_default: true,
-        },
-        super::AudioHostSummary {
-            id: String::from("asio"),
-            label: String::from("ASIO"),
-            is_default: false,
-        },
-    ];
-
-    state.audio_backend_dropdown_open = false;
-    let closed =
-        radiant::runtime::UiSurface::new(super::audio_settings_popover(&state).into_node()).frame(
-            Rect::from_min_size(Point::new(0.0, 0.0), Vector2::new(480.0, 360.0)),
-            &radiant::theme::ThemeTokens::default(),
-        );
-    state.audio_backend_dropdown_open = true;
-    let open = radiant::runtime::UiSurface::new(super::audio_settings_popover(&state).into_node())
-        .frame(
-            Rect::from_min_size(Point::new(0.0, 0.0), Vector2::new(480.0, 360.0)),
-            &radiant::theme::ThemeTokens::default(),
-        );
-
-    let text_top = |frame: &radiant::runtime::SurfaceFrame, label: &str| {
-        frame
-            .paint_plan
-            .primitives
-            .iter()
-            .find_map(|primitive| match primitive {
-                PaintPrimitive::Text(text) if text.text.as_str() == label => Some(text.rect.min.y),
-                _ => None,
-            })
-            .unwrap_or_else(|| panic!("expected text {label}"))
-    };
-    let text_index = |frame: &radiant::runtime::SurfaceFrame, label: &str| {
-        frame
-            .paint_plan
-            .primitives
-            .iter()
-            .position(|primitive| match primitive {
-                PaintPrimitive::Text(text) => text.text.as_str() == label,
-                _ => false,
-            })
-            .unwrap_or_else(|| panic!("expected text {label}"))
-    };
-
-    assert_eq!(text_top(&closed, "Output"), text_top(&open, "Output"));
-    assert_eq!(
-        text_top(&closed, "Sample Rate"),
-        text_top(&open, "Sample Rate")
-    );
-    assert!(text_top(&open, "WASAPI (default)") > text_top(&open, "Output"));
-    assert!(text_index(&open, "WASAPI (default)") > text_index(&open, "Output"));
-}
-
-#[test]
-fn audio_backend_dropdown_toggle_and_close_are_ui_only() {
-    let mut state = gui_state_for_span_tests();
-
-    state.apply_message(
-        super::GuiMessage::ToggleAudioBackendDropdown,
-        &mut ui::UpdateContext::default(),
-    );
-    assert!(state.audio_backend_dropdown_open);
-
-    state.apply_message(
-        super::GuiMessage::CloseAudioSettingsDropdowns,
-        &mut ui::UpdateContext::default(),
-    );
-    assert!(!state.audio_backend_dropdown_open);
-
-    state.apply_message(
-        super::GuiMessage::ToggleAudioBackendDropdown,
-        &mut ui::UpdateContext::default(),
-    );
-    assert!(state.audio_backend_dropdown_open);
-
-    state.apply_message(
-        super::GuiMessage::ToggleAudioOutputDropdown,
-        &mut ui::UpdateContext::default(),
-    );
-    assert!(!state.audio_backend_dropdown_open);
-    assert!(state.audio_output_dropdown_open);
-
-    state.apply_message(
-        super::GuiMessage::ToggleAudioSampleRateDropdown,
-        &mut ui::UpdateContext::default(),
-    );
-    assert!(!state.audio_output_dropdown_open);
-    assert!(state.audio_sample_rate_dropdown_open);
-
-    state.apply_message(
-        super::GuiMessage::CloseAudioSettingsDropdowns,
-        &mut ui::UpdateContext::default(),
-    );
-    assert!(!state.audio_sample_rate_dropdown_open);
-
-    state.apply_message(
-        super::GuiMessage::CloseAudioSettings,
-        &mut ui::UpdateContext::default(),
-    );
-    assert!(!state.audio_backend_dropdown_open);
-    assert!(!state.audio_output_dropdown_open);
-    assert!(!state.audio_sample_rate_dropdown_open);
 }
 
 #[test]
